pydough-analytics/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydough-analytics"
version = "0.1.0"
description = "Text-to-analytics toolkit built on PyDough"
readme = "README.md"
authors = [{name = "Bodo.ai"}]
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "pydough>=0.9.0",
    "pandas>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.3.0",
    "pydantic>=2.8.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "inflect>=7.0.0"
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
mysql = ["mysql-connector-python>=8.0.0"]
snowflake = ["snowflake-connector-python[pandas]>=3.0.0"]
mcp = ["fastmcp>=2.12.0"]

[project.scripts]
pydough-analytics = "pydough_analytics.cli:app"
pydough-analytics-mcp = "pydough_analytics.mcp.server:main"

[tool.typer]
auto_envvar_prefix = "PYDOUGH_ANALYTICS"


[tool.pytest.ini_options]
markers = [
    "live_llm: marks tests that call the live Gemini API (off by default)",
]


[tool.hatch.build.targets.sdist]
include = ["src/pydough_analytics", "README.md", "LICENSE"]

[tool.hatch.build.targets.wheel]
packages = ["src/pydough_analytics"]
